+++ src/rovr/screens/zd_to_directory.py
@@ -1,24 +1,21 @@
 import asyncio
 import contextlib
+from typing import cast
 
-from textual import events, work
+from textual import on, work
 from textual.app import ComposeResult
 from textual.containers import VerticalGroup
-from textual.screen import ModalScreen
 from textual.widgets import Input, OptionList
 from textual.worker import WorkerCancelled
 
 from rovr.classes.textual_options import ModalSearcherOption
-from rovr.components import DoubleClickableOptionList
-from rovr.functions.utils import check_key, should_cancel
+from rovr.components import DoubleClickableOptionList, ModalSearchScreen
+from rovr.functions.utils import should_cancel
 from rovr.variables.constants import config
 
 
-class ZDToDirectory(ModalScreen):
+class ZDToDirectory(ModalSearchScreen):
     """Screen with a dialog to z to a directory, using zoxide"""
-
-    def __init__(self, **kwargs) -> None:
-        super().__init__(**kwargs)
 
     def compose(self) -> ComposeResult:
         with VerticalGroup(id="zoxide_group", classes="zoxide_group"):
@@ -33,15 +30,10 @@ class ZDToDirectory(ModalScreen):
             )
 
     def on_mount(self) -> None:
-        self.zoxide_input: Input = self.query_one("#zoxide_input")
-        self.zoxide_input.border_title = "zoxide"
-        self.zoxide_input.focus()
-        self.zoxide_options: DoubleClickableOptionList = self.query_one(
-            "#zoxide_options"
-        )
-        self.zoxide_options.border_title = "Folders"
-        self.zoxide_options.can_focus = False
-        self.zoxide_updater(Input.Changed(self.zoxide_input, value=""))
+        super().on_mount()
+        self.search_input.border_title = "zoxide"
+        self.search_options.border_title = "Folders"
+        self.zoxide_updater(Input.Changed(self.search_input, value=""))
 
     def on_input_changed(self, event: Input.Changed) -> None:
         self.zoxide_updater(event=event)
@@ -87,6 +79,7 @@ class ZDToDirectory(ModalScreen):
 
         zoxide_cmd += search_term.split()
 
+        zoxide_process = None
         try:
             zoxide_process = await asyncio.create_subprocess_exec(
                 *zoxide_cmd,
@@ -95,15 +88,11 @@ class ZDToDirectory(ModalScreen):
             )
             stdout, _ = await asyncio.wait_for(zoxide_process.communicate(), timeout=3)
         except (OSError, asyncio.exceptions.TimeoutError) as exc:
-            if isinstance(exc, asyncio.exceptions.TimeoutError):
+            if isinstance(exc, asyncio.exceptions.TimeoutError) and zoxide_process:
                 zoxide_process.kill()
-                with contextlib.suppress(
-                    asyncio.exceptions.TimeoutError, ProcessLookupError
-                ):
-                    await asyncio.wait_for(zoxide_process.wait(), timeout=1)
             # zoxide not installed
-            self.zoxide_options.clear_options()
-            self.zoxide_options.add_option(
+            self.search_options.clear_options()
+            self.search_options.add_option(
                 ModalSearcherOption(
                     None,
                     "  zoxide is missing on $PATH or cannot be executed"
@@ -113,6 +102,7 @@ class ZDToDirectory(ModalScreen):
                 )
             )
             return
+
         # check 2 for queue, to ignore mounting as a whole
         if should_cancel():
             return
@@ -120,15 +110,20 @@ class ZDToDirectory(ModalScreen):
             stdout = stdout.decode()
             worker = self.create_options(show_scores, stdout)
             try:
-                options: list[ModalSearcherOption] = await worker.wait()
+                options_result = await worker.wait()
+                if options_result is not None:
+                    options = options_result
+                else:
+                    return
             except WorkerCancelled:
                 return  # anyways
             if options is None:
                 return
-            if len(options) == len(self.zoxide_options.options) and all(
+            if len(options) == len(self.search_options.options) and all(
                 isinstance(options[i], ModalSearcherOption)
-                and isinstance(self.zoxide_options.options[i], ModalSearcherOption)
-                and options[i].file_path == self.zoxide_options.options[i].file_path
+                and isinstance(self.search_options.options[i], ModalSearcherOption)
+                and options[i].file_path
+                == cast(ModalSearcherOption, self.search_options.options[i]).file_path
                 for i in range(len(options))
             ):  # ie same~ish query, resulting in same result
                 pass
@@ -138,43 +133,29 @@ class ZDToDirectory(ModalScreen):
                 # raised, or just nothing showing up. By having the clear
                 # options and add options functions nearby, it hopefully
                 # reduces the likelihood of an empty option list
-                self.zoxide_options.set_options(options)
-                self.zoxide_options.remove_class("empty")
-                self.zoxide_options.highlighted = 0
+                self.search_options.set_options(options)
+                self.search_options.remove_class("empty")
+                self.search_options.highlighted = 0
                 if should_cancel():
                     return
         else:
             # No Matches to the query text
-            self.zoxide_options.clear_options()
-            self.zoxide_options.add_option(
+            self.search_options.clear_options()
+            self.search_options.add_option(
                 ModalSearcherOption(None, "  --No matches found--", disabled=True),
             )
-            self.zoxide_options.add_class("empty")
-            self.zoxide_options.border_subtitle = "0/0"
+            self.search_options.add_class("empty")
+            self.search_options.border_subtitle = "0/0"
 
-    def on_input_submitted(self, event: Input.Submitted) -> None:
-        if any(
-            worker.is_running and worker.node is self for worker in self.app.workers
-        ):
-            return
-        if self.zoxide_options.highlighted is None:
-            self.zoxide_options.highlighted = 0
-        if (
-            self.zoxide_options.highlighted_option
-            and self.zoxide_options.highlighted_option.disabled
-        ):
-            return
-        self.zoxide_options.action_select()
-
-    # You can't manually tab into the option list, but you can click, so I guess
     @work(exclusive=True)
-    async def on_option_list_option_selected(
-        self, event: OptionList.OptionSelected
-    ) -> None:
+    @on(OptionList.OptionSelected)
+    async def handle_zd_option_selected(self, event: OptionList.OptionSelected) -> None:
+        event.stop()
         if not isinstance(event.option, ModalSearcherOption):
             # theoretically this shouldnt happen, but precautions
             self.dismiss(None)
             return
+
         selected_value = event.option.file_path
         if selected_value is None:
             self.dismiss(None)
@@ -193,37 +174,6 @@ class ZDToDirectory(ModalScreen):
             self.dismiss(selected_value)
         else:
             self.dismiss(None)
-
-    def on_key(self, event: events.Key) -> None:
-        if check_key(event, config["keybinds"]["filter_modal"]["exit"]):
-            event.stop()
-            self.dismiss(None)
-        elif check_key(event, config["keybinds"]["filter_modal"]["down"]):
-            event.stop()
-            if self.zoxide_options.options:
-                self.zoxide_options.action_cursor_down()
-        elif check_key(event, config["keybinds"]["filter_modal"]["up"]):
-            event.stop()
-            if self.zoxide_options.options:
-                self.zoxide_options.action_cursor_up()
-        elif event.key == "tab":
-            event.stop()
-            self.focus_next()
-        elif event.key == "shift+tab":
-            event.stop()
-            self.focus_previous()
-
-    def on_option_list_option_highlighted(
-        self, event: OptionList.OptionHighlighted
-    ) -> None:
-        if (
-            self.zoxide_options.highlighted is None
-            or self.zoxide_options.option_count == 0
-            or self.zoxide_options.get_option_at_index(0).disabled
-        ):
-            self.zoxide_options.border_subtitle = "0/0"
-        else:
-            self.zoxide_options.border_subtitle = f"{self.zoxide_options.highlighted + 1}/{self.zoxide_options.option_count}"
 
     @work(thread=True)
     def create_options(
@@ -248,9 +198,3 @@ class ZDToDirectory(ModalScreen):
             if should_cancel():
                 return
         return options
-
-    def on_click(self, event: events.Click) -> None:
-        if event.widget is self:
-            # ie click outside
-            event.stop()
-            self.dismiss(None)
